Remove deleteMyPoll, move logic into deletePoll
Admins can delete anyones poll, users can delete their own polls

+++ backend/src/main/java/com/hvl/feedApp/controller/PollController.java
@@ -7,6 +7,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.hvl.feedApp.Agent;
+import com.hvl.feedApp.Enums.Role;
 import com.hvl.feedApp.Poll;
 import com.hvl.feedApp.Vote;
 import com.hvl.feedApp.security.Authenticator;
@@ -107,30 +108,17 @@ public class PollController {
         }
         throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid user credentials");
     }
-    @DeleteMapping(path = "/deleteMyPoll/{pollID}")
-    public String deleteMyPoll(@RequestHeader(HttpHeaders.AUTHORIZATION) String bAuth, @PathVariable("pollID") Long pollID){
-        if (authenticator.isAuthenticated(bAuth)) {
-            if (authorizer.isAuthorized(authenticator.getUser(), "/deleteMyPoll/{pollID}", "DELETE")) {
-                try{
-                    Poll poll = pollService.getPollById(pollID);
-                    if ( (poll.getOwner()) != (authenticator.getUser())){
-                        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Can't delete another users Poll");
-                    }
-                    pollService.deletePoll(pollID);
-                    pollService.getPollById(pollID);
-                    throw new ResponseStatusException(HttpStatus.CONFLICT, "Poll with ID "+pollID+" was not successfully deleted.");
-                } catch (IllegalStateException e) {
-                    return "Successfully deleted Poll with ID "+pollID;
-                }
-            }
-        }
-        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid user credentials");
-    }
     @DeleteMapping(path = "{pollID}")
     public String deletePoll(@RequestHeader(HttpHeaders.AUTHORIZATION) String bAuth, @PathVariable("pollID") Long pollID){
         if (authenticator.isAuthenticated(bAuth)) {
             if (authorizer.isAuthorized(authenticator.getUser(), "/polls/{pollID}", "DELETE")) {
                 try{
+                    Poll poll = pollService.getPollById(pollID);
+                    Boolean userIsAdmin = authenticator.getUser().getRole() == Role.ADMIN;
+                    Boolean userOwnsPoll = poll.getOwner() == authenticator.getUser();
+                    if (!userIsAdmin && !userOwnsPoll) {
+                        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Can't delete another users Poll");
+                    }
                     pollService.deletePoll(pollID);
                     pollService.getPollById(pollID);
                     throw new ResponseStatusException(HttpStatus.CONFLICT, "Poll with ID "+pollID+" was not successfully deleted.");
